refactor: network層のリファクタリング
- ファイル切り分け
- lintエラー修正

Fixed #47
